REFACTOR-002: extract pack invariants into ./parse/packs.ts
Pure structural move, no behavior change. Moves checkMetadata,
checkPackDependencyList, checkPackQualification, and checkAddedPacks
(module-private, called from checkProjectIntegrity/parseProjectMetadata)
out of parse.ts into their own module.

Fourth of the REFACTOR-002 stack; builds on the automation-invariants
extraction.

Refs #143
